Alert dialogs while deleting something delete comments when post is deleted

+++ app/src/main/java/com/project/findme/mainactivity/mainfragments/ui/comment/CommentFragment.kt
@@ -3,12 +3,14 @@ package com.project.findme.mainactivity.mainfragments.ui.comment
 import android.os.Bundle
 import android.util.Log
 import android.view.View
+import androidx.appcompat.app.AlertDialog
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.project.findme.adapter.CommentAdapter
+import com.project.findme.data.entity.Comment
 import com.project.findme.utils.EventObserver
 import com.project.findme.utils.snackbar
 import com.ryan.findme.R
@@ -17,7 +19,7 @@ import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 
 @AndroidEntryPoint
-class CommentFragment: Fragment(R.layout.fragment_comment) {
+class CommentFragment : Fragment(R.layout.fragment_comment) {
 
     @Inject
     lateinit var commentAdapter: CommentAdapter
@@ -39,17 +41,31 @@ class CommentFragment: Fragment(R.layout.fragment_comment) {
         subscribeToObserve()
     }
 
+    private fun showConfirmationDialog(comment: Comment) {
+        AlertDialog.Builder(requireContext())
+            .setTitle("Delete Comment")
+            .setMessage("Are you sure you want to delete this comment?")
+            .setPositiveButton(
+                "Yes"
+            ) { _, _ ->
+                viewModel.deleteComment(comment)
+            }
+            .setNegativeButton("No", null)
+            .setIcon(android.R.drawable.ic_dialog_alert)
+            .show()
+    }
+
     private fun setupRecyclerView() {
         binding.commentRv.apply {
             layoutManager = LinearLayoutManager(requireContext())
             adapter = commentAdapter
         }
         commentAdapter.setDeleteListener { comment ->
-            viewModel.deleteComment(comment)
+            showConfirmationDialog(comment)
         }
     }
 
-     private fun subscribeToObserve() {
+    private fun subscribeToObserve() {
         viewModel.createCommentStatus.observe(viewLifecycleOwner, EventObserver(
             onError = { error ->
                 binding.apply {
@@ -64,54 +80,54 @@ class CommentFragment: Fragment(R.layout.fragment_comment) {
                     sendCommentBt.isClickable = false
                 }
             }
-        ){
+        ) {
             binding.apply {
                 commentPb.isVisible = false
                 sendCommentBt.isClickable = true
                 commentTextInputEt.setText("")
                 commentAdapter.comment += it
-                Log.d("CommentFragment: ",it.toString())
+                Log.d("CommentFragment: ", it.toString())
             }
         })
 
-         viewModel.getCommentsForPostStatus.observe(viewLifecycleOwner, EventObserver(
-             onError = { error ->
-                 snackbar(error)
-                 Log.d("CommentFragment: ", error)
-                 binding.apply {
-                     commentPb.isVisible = false
-                     sendCommentBt.isClickable = true
-                 }
-             },
-             onLoading = {
-                 binding.apply {
-                     commentPb.isVisible = true
-                     sendCommentBt.isClickable = false
-                     commentAdapter.comment = listOf()
-                 }
-             }
-         ){ commentList ->
-             binding.apply {
-                 commentPb.isVisible = false
-                 sendCommentBt.isClickable = true
-                 Log.d("CommentFragment: ", commentList.toString())
-                 commentAdapter.comment = commentList
-             }
-         })
+        viewModel.getCommentsForPostStatus.observe(viewLifecycleOwner, EventObserver(
+            onError = { error ->
+                snackbar(error)
+                Log.d("CommentFragment: ", error)
+                binding.apply {
+                    commentPb.isVisible = false
+                    sendCommentBt.isClickable = true
+                }
+            },
+            onLoading = {
+                binding.apply {
+                    commentPb.isVisible = true
+                    sendCommentBt.isClickable = false
+                    commentAdapter.comment = listOf()
+                }
+            }
+        ) { commentList ->
+            binding.apply {
+                commentPb.isVisible = false
+                sendCommentBt.isClickable = true
+                Log.d("CommentFragment: ", commentList.toString())
+                commentAdapter.comment = commentList
+            }
+        })
 
-         viewModel.deleteCommentStatus.observe(viewLifecycleOwner, EventObserver(
-             onError = {
-                 snackbar(it)
-                 binding.apply {
-                     commentPb.isVisible = false
-                 }
-             },
-             onLoading = {
+        viewModel.deleteCommentStatus.observe(viewLifecycleOwner, EventObserver(
+            onError = {
+                snackbar(it)
+                binding.apply {
+                    commentPb.isVisible = false
+                }
+            },
+            onLoading = {
                 binding.commentPb.isVisible = true
-             }
-         ){
-             commentAdapter.comment -= it
-             binding.commentPb.isVisible = false
-         })
-     }
+            }
+        ) {
+            commentAdapter.comment -= it
+            binding.commentPb.isVisible = false
+        })
+    }
 }
